Tests/LibWeb/Text/expected/SVG/svg-transform-list-indexed-setter.txt

baseVal === animVal: false
initial length: 0
transform type: 0
new index: NoModificationAllowedError: Cannot modify a read-only list
wrong type: TypeError: Value must be an SVGTransform
